# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediNear API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops:
    get:
      operationId: listShops
      tags: [shops]
      summary: List all medicine shops
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: isOpen
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of shops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shop"
    post:
      operationId: createShop
      tags: [shops]
      summary: Create a new medicine shop (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopInput"
      responses:
        "201":
          description: Created shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /shops/{id}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get a shop by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "404":
          description: Not found
    patch:
      operationId: updateShop
      tags: [shops]
      summary: Update a shop (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopUpdate"
      responses:
        "200":
          description: Updated shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    delete:
      operationId: deleteShop
      tags: [shops]
      summary: Delete a shop (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /lab-tests:
    get:
      operationId: listLabTests
      tags: [lab-tests]
      summary: List lab test categories and bookings
      responses:
        "200":
          description: List of lab tests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabTest"
    post:
      operationId: createLabTest
      tags: [lab-tests]
      summary: Book a lab test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabTestInput"
      responses:
        "201":
          description: Created lab test booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabTest"
  /lab-tests/{id}:
    get:
      operationId: getLabTest
      tags: [lab-tests]
      summary: Get a lab test booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lab test details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabTest"
    patch:
      operationId: updateLabTest
      tags: [lab-tests]
      summary: Update lab test (admin — add report URL)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabTestUpdate"
      responses:
        "200":
          description: Updated lab test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabTest"
    delete:
      operationId: deleteLabTest
      tags: [lab-tests]
      summary: Delete a lab test booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a doctor or shop
      parameters:
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: shopId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Shop:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        isOpen:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        openingTime:
          type: ["string", "null"]
        closingTime:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, address, phone, isOpen, reviewCount, createdAt]
    ShopInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        phone:
          type: string
        isOpen:
          type: boolean
        imageUrl:
          type: string
        lat:
          type: number
        lng:
          type: number
        openingTime:
          type: string
        closingTime:
          type: string
      required: [name, address, phone]
    ShopUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        isOpen:
          type: boolean
        imageUrl:
          type: string
        lat:
          type: number
        lng:
          type: number
        openingTime:
          type: string
        closingTime:
          type: string
...
    LabTest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        testName:
          type: string
        category:
          type: string
        shopId:
          type: integer
        scheduledDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, completed, cancelled]
        reportUrl:
          type: ["string", "null"]
        price:
          type: number
        shopName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, testName, category, shopId, status, price, createdAt]
    LabTestInput:
      type: object
      properties:
        testName:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        shopId:
          type: integer
        scheduledDate:
          type: string
        price:
          type: number
      required: [testName, category, shopId, price]
    LabTestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, completed, cancelled]
        reportUrl:
          type: string
        scheduledDate:
          type: string
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        doctorId:
          type: ["integer", "null"]
        shopId:
          type: ["integer", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, rating, createdAt]
    ReviewInput:
      type: object
      properties:
        doctorId:
          type: integer
        shopId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [rating]
...