# lib/api-zod/src/generated/types/labTestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediNear API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LabTestInput {
  /** @minLength 1 */
  testName: string;
  /** @minLength 1 */
  category: string;
  shopId: number;
  scheduledDate?: string;
  price: number;
}