# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediNear API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  userId: string;
  /** @nullable */
  doctorId?: number | null;
  /** @nullable */
  shopId?: number | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
  /** @nullable */
  userName?: string | null;
  createdAt: string;
}