# lib/api-zod/src/generated/types/shopUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediNear API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShopUpdate {
  name?: string;
  address?: string;
  phone?: string;
  isOpen?: boolean;
  imageUrl?: string;
  lat?: number;
  lng?: number;
  openingTime?: string;
  closingTime?: string;
}